mv storage miner api to separate file (#393)
