+++ core/const.hpp
@@ -61,7 +61,6 @@ namespace fc {
 
   extern EpochDuration kBreezeGasTampingDuration;
 
-  constexpr uint64_t kMinerApiVersion{0};
   constexpr uint64_t kDefaultStorageWeight{10};
 
   extern EpochDuration kInteractivePoRepConfidence;
