+++ core/miner/main/main.cpp
@@ -15,7 +15,7 @@
 #include "api/rpc/info.hpp"
 #include "api/rpc/make.hpp"
 #include "api/rpc/ws.hpp"
-#include "api/storage_api.hpp"
+#include "api/storage_miner/storage_api.hpp"
 #include "clock/impl/utc_clock_impl.hpp"
 #include "codec/json/json.hpp"
 #include "common/file.hpp"
@@ -36,7 +36,6 @@
 #include "proofs/proof_param_provider.hpp"
 #include "sector_storage/fetch_handler.hpp"
 #include "sector_storage/impl/manager_impl.hpp"
-#include "sector_storage/impl/remote_worker.hpp"
 #include "sector_storage/impl/scheduler_impl.hpp"
 #include "sector_storage/stores/impl/index_impl.hpp"
 #include "sector_storage/stores/impl/local_store.hpp"
@@ -53,21 +52,12 @@
 #include "vm/actor/builtin/v0/storage_power/storage_power_actor_export.hpp"
 
 namespace fc {
-  using api::RegisteredSealProof;
-  using api::SignedStorageAsk;
   using boost::asio::io_context;
   using common::span::cbytes;
   using config::configProfile;
   using libp2p::multi::Multiaddress;
-  using markets::retrieval::RetrievalAsk;
-  using primitives::StorageID;
   using primitives::address::Address;
   using primitives::sector::RegisteredSealProof;
-  using primitives::sector_file::SectorFileType;
-  using sector_storage::RemoteWorker;
-  using sector_storage::stores::FsStat;
-  using sector_storage::stores::HealthReport;
-  using sector_storage::stores::StorageInfo;
   using storage::BufferMap;
   namespace uuids = boost::uuids;
 
@@ -421,100 +411,16 @@ namespace fc {
             miner)};
     retrieval_provider->start();
 
-    auto mapi{std::make_shared<api::StorageMinerApi>()};
-    mapi->ActorAddress = [&]() { return miner->getAddress(); };
-
-    mapi->ActorSectorSize =
-        [&](Address addr) -> outcome::result<api::SectorSize> {
-      OUTCOME_TRY(miner_info, napi->StateMinerInfo(addr, {}));
-      return miner_info.sector_size;
-    };
-
-    mapi->DealsImportData = [&](auto &proposal, auto &path) {
-      return storage_provider->importDataForDeal(proposal, path);
-    };
-    mapi->MarketGetAsk = [&]() -> outcome::result<SignedStorageAsk> {
-      return stored_ask->getAsk(*config.actor);
-    };
-    mapi->MarketGetRetrievalAsk = [&]() -> outcome::result<RetrievalAsk> {
-      return retrieval_provider->getAsk();
-    };
-    mapi->MarketSetAsk = [&](auto &price,
-                             auto &verified_price,
-                             auto duration,
-                             auto min_piece_size,
-                             auto max_piece_size) -> outcome::result<void> {
-      return stored_ask->addAsk(
-          {
-              .price = price,
-              .verified_price = verified_price,
-              .min_piece_size = min_piece_size,
-              .max_piece_size = max_piece_size,
-              .miner = *config.actor,
-          },
-          duration);
-    };
-    mapi->MarketSetRetrievalAsk = [&](auto &ask) -> outcome::result<void> {
-      retrieval_provider->setAsk(ask);
-      return outcome::success();
-    };
-    mapi->PledgeSector = [&]() -> outcome::result<void> {
-      return sealing->pledgeSector();
-    };
-    mapi->Version = [] {
-      return api::VersionResult{"fuhon-miner", kMinerApiVersion, 0};
-    };
-
-    // TODO(ortyomka): [FIL-347] remove it
-    mapi->SealProof = [&] { return wscheduler->getSealProofType(); };
-
-    mapi->StorageBestAlloc = [&](const SectorFileType &allocate,
-                                 RegisteredSealProof seal_proof_type,
-                                 bool sealing_mode) {
-      return sector_index->storageBestAlloc(
-          allocate, seal_proof_type, sealing_mode);
-    };
-    mapi->StorageFindSector =
-        [&](const SectorId &sector,
-            const SectorFileType &file_type,
-            boost::optional<RegisteredSealProof> fetch_seal_proof_type) {
-          return sector_index->storageFindSector(
-              sector, file_type, fetch_seal_proof_type);
-        };
-    mapi->StorageDropSector = [&](const StorageID &storage_id,
-                                  const SectorId &sector,
-                                  const SectorFileType &file_type) {
-      return sector_index->storageDropSector(storage_id, sector, file_type);
-    };
-    mapi->StorageDeclareSector = [&](const StorageID &storage_id,
-                                     const SectorId &sector,
-                                     const SectorFileType &file_type,
-                                     bool primary) {
-      return sector_index->storageDeclareSector(
-          storage_id, sector, file_type, primary);
-    };
-    mapi->StorageReportHealth = [&](const StorageID &storage_id,
-                                    const HealthReport &report) {
-      return sector_index->storageReportHealth(storage_id, report);
-    };
-    mapi->StorageAttach = [&](const StorageInfo &storage_info,
-                              const FsStat &stat) {
-      return sector_index->storageAttach(storage_info, stat);
-    };
-    mapi->StorageInfo = [&](const StorageID &storage_id) {
-      return sector_index->getStorageInfo(storage_id);
-    };
-
-    mapi->WorkerConnect =
-        [&](const std::string &address) -> outcome::result<void> {
-      OUTCOME_TRY(maddress, libp2p::multi::Multiaddress::create(address));
-      OUTCOME_TRY(worker,
-                  RemoteWorker::connectRemoteWorker(*io, mapi, maddress));
-
-      spdlog::info("Connected to a remote worker at {}", address);
-
-      return manager->addWorker(std::move(worker));
-    };
+    auto mapi = api::makeStorageApi(io,
+                                    napi,
+                                    *config.actor,
+                                    miner,
+                                    sector_index,
+                                    manager,
+                                    wscheduler,
+                                    stored_ask,
+                                    storage_provider,
+                                    retrieval_provider);
 
     std::map<std::string, std::shared_ptr<api::Rpc>> mrpc;
     mrpc.emplace("/rpc/v0", api::makeRpc(*mapi));
