+++ core/sector_storage/impl/manager_impl.hpp
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef CPP_FILECOIN_CORE_SECTOR_STORAGE_IMPL_MANAGER_IMPL_HPP
-#define CPP_FILECOIN_CORE_SECTOR_STORAGE_IMPL_MANAGER_IMPL_HPP
+#pragma once
 
 #include "sector_storage/manager.hpp"
 
@@ -151,5 +150,3 @@ namespace fc::sector_storage {
   };
 
 }  // namespace fc::sector_storage
-
-#endif  // CPP_FILECOIN_CORE_SECTOR_STORAGE_IMPL_MANAGER_IMPL_HPP
