+++ core/sector_storage/scheduler.hpp
@@ -3,8 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-#ifndef CPP_FILECOIN_CORE_SECTOR_STORAGE_SCHEDULER_HPP
-#define CPP_FILECOIN_CORE_SECTOR_STORAGE_SCHEDULER_HPP
+#pragma once
 
 #include "common/outcome.hpp"
 #include "primitives/seal_tasks/task.hpp"
@@ -44,5 +43,3 @@ namespace fc::sector_storage {
 }  // namespace fc::sector_storage
 
 OUTCOME_HPP_DECLARE_ERROR(fc::sector_storage, SchedulerErrors);
-
-#endif  // CPP_FILECOIN_CORE_SECTOR_STORAGE_SCHEDULER_HPP
